api/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
asyncpg==0.29.0
alembic==1.13.2
pydantic==2.8.2
pydantic-settings==2.3.4
httpx==0.27.0
python-multipart==0.0.22
aiosqlite==0.20.0
better-profanity==0.7.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
jinja2==3.1.4
aiosmtplib==3.0.1
